moved accordion utils into step2 utils
